Add files via upload
Project Overview:

In this project, we will create a Virtual Network (VNet) in Azure and deploy an Azure Firewall to control and secure inbound and outbound traffic.

Behind the firewall, we will configure a subnet called Web Application Subnet, in which we will deploy a Virtual Machine (VM). On this VM, we will install Nginx and host a simple HTML web page.

Since the VM will have only a private IP address, it will not be directly accessible from the internet. To securely connect to the VM, we will use Azure Bastion. This service allows secure RDP/SSH connections to the VM through the Azure Portal without exposing the VM to the public internet.

We will also configure firewall rules so that authenticated users can bypass the firewall and access the Nginx application hosted on the VM.